requirements.txt
opencv-python
mediapipe
numpy
pygame
Pillow